nvme: add doorbell buffer config support
Add support for setting up doorbell buffers (shadow doorbells). By
default, this is enabled for the admin queue as well. However, some
devices may not implement this (in violation of the spec), so introduce
a quirk for such devices. For now, this quick must be applied manually
when intializing the driver.

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>