rq: verify number of prps
Change the APIs for mapping buffers into command PRPs. This allows the
mapping functions to error out if the buffer(s) cannot be mapped into
the request for some reason (like using too many PRPs or if alignment
requiremens are violated).

Until now, some of this has resulted in assertions or for the "too many
PRPs" case, simply ignored.

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>