vfio/iommu: Move the early return up in iommu conf
If the iommu is already configured, the container iommu configuration is
skipped with an early return. This usually happened after setting the
vfio iommu type. We move the early return in such a way that we can use
it for when we deal with two APIs. Having it early in the
vfio_container_configure_iommu function means that we can use it for
both vfio and iommu user APIs.

Signed-off-by: Joel Granados <j.granados@samsung.com>