vfio/iommu: add new func to only configure iommu
This is a preparation commit that sets things up to introduce two ways
of configuring the iommu user API. In this commit we move all logic from
vfio_group_set_container to vfio_get_group_fd and add logic to the later
to undo things in case of error. Having all that logic in
vfio_get_group_fd allows us to remove vfio_group_set_container
completely (it was not doing much to begin with) and replace it with
vfio_configure_iommu.

There is no functional change in vfio_configure_iommu but now it can
easily be modified to multiplex the two ways of configuring iommu user
APIs.

Signed-off-by: Joel Granados <j.granados@samsung.com>