vfio: remove ephemeral mappings
Until now we have used an allocation strategy similar to that from the
vfio helpers in QEMU. That is, we allocate "sticky" or "fixed" iovas
from low addresses in the iova space and "ephemeral" or "temporary"
iovas from high addresses. This works well, but they are rarely used and
as we move towards an iommufd-first backend we will start to use
automatic iova allocation through the kernel allocator.

Drop the ephemeral mapping functionality and instead reserve the low 64k
of the iova address space for internal use.

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>