iommu: re-add support for ephemeral mappings
The nvme_sync() helper function maps a buffer, issues the command, and
unmaps the buffer. When we removed ephemeral IOVAs we relied on the user
to handle these temporary mappings from a reserved set of IOVAs at the
start of the IOVA space. While ephemeral IOVAs are used rarely, it is
nicer to be able to just pass a buffer that we know will only be mapped
temporarily. The new iommufd backend does not care about temporary
mappings since it uses the kernel iova allocation that handles this just
fine.

The vfio backend still needs some kind of special treatment. We retain a
reserved space (by default, 64k) at the start of the iova range and use
that for temporary (ephemeral) mappings, only now, instead of requiring
the user to map from that area explicitly, a new IOMMU_MAP_EPHEMERAL
flags may be passed to indicate the temporary nature of the iova. Like
before, we recycle the whole ephemeral space when no mappings are in
use.

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>