build: add a meson build option for linux headers
One might need to point the build to a linux kernel include directory
different from the one that is currently installed. To avoid having to
constantly install the linux kernel headers on the system, add a meson
build option that adds a user defined path to the include paths of the
project.

Signed-off-by: Joel Granados <j.granados@samsung.com>
[k.jensen: refactor]
Signed-off-by: Klaus Jensen <k.jensen@samsung.com>