iommu: Refactor and introduce vfio_get_group helper
Previous pci_get_iommu_group() returns a string path of vfio group.  The
string path is actually required for nothing but getting the group fd.
The actual instance for the vfio group is (struct vfio_group *) so that
this patch introduces a new helper vfio_get_group() returning the vfio
group instance.

This is a prep patch to introduce a helper to get the vfio group
instance without the path string value.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>