iommufd: remove put_device_fd logic
An IOAS may have several devices attached, so devfd does not belong as a
member within struct iommu_ioas. Remove it.

This member was used in put_device_fd, but the iommufd backend does not
need such cleanup. It is enough to close the device file descriptor
(which we do in vfio_pci_close).

Signed-off-by: Klaus Jensen <k.jensen@samsung.com>