nvme/core: abstract mmio
The mmio helpers assume that we can read and write to a specific
address. The proposed DriverKit backend cannot do that and must call
specific functions on a base address and an offset.

In the core, prepare for DriverKit to provide its own implementation of
doing that.

This is a simplified version of commit "mmio: convert calls of 'addr +
offset' to 'base-addr, offset'" to limit the change to the nvme core.

Signed-off-by: Mads Ynddal <m.ynddal@samsung.com>
[k.jensen: limit the change to nvme/core for now]
Signed-off-by: Klaus Jensen <k.jensen@samsung.com>