build: add kernel headers in the flake package
Having the Linux headers available is necessary to activate the IOMMUFD
path. This patch does not activate the path by default as nixos 24.11
still has a kernel that is too old (6.10), but it adds the reference in
the flake so it is ready for when the next stable nixos version comes
out.

Signed-off-by: Joel Granados <j.granados@samsung.com>