iommu/vfio: introduce best-fit iova allocation
Introduce a best-fit IOVA allocation policy to vfio_iommu_type1 device
to reuse previously allocated and freed IOVA ranges in the next
allocation.  Unlike `iommufd` which allocates the IOVA address in the
kernel side, `vfio_iommu_type1` device should manage the IOVA ranges in
libvfn to avoid permanently incremental memory allocation for a
long-lived application like ``unvme-cli``.

Firstly try to allocate from @free_iovas skiplist for a new IOVA
allocation with a best-fit allocation method, and if not, it goes to the
previous way with `__iova_reserve()`.  Other than this, no functional
changes.

Signed-off-by: Minwoo Im <minwoo.im@samsung.com>